requirements.txt
Flask
Flask-Login
Flask-Migrate
Flask-SQLAlchemy
PyMySQL
python-dotenv